split out rank exit event into its own file
